cost per scale point is now a double

+++ src/structs.h
@@ -3904,7 +3904,7 @@ struct ability_data {
 	bitvector_t targets;	// ATAR_ flags
 	int cost_type;	// HEALTH, MANA, etc.
 	int cost;	// amount of h/v/m
-	int cost_per_scale_point;	// cost modifier when scaled
+	double cost_per_scale_point;	// cost modifier when scaled
 	struct resource_data *resource_cost;	// additional costs
 	any_vnum cooldown;	// generic cooldown if any
 	int cooldown_secs;	// how long to cooldown, if any
